hanbun/pyproject.toml
[project]
name = "hanbun"
version = "0.1.0"
description = "Analyze Korean text against vocabulary and frequency lists"
readme = "README.md"
authors = [
    { name = "Larry Grullon-Polanco", email = "larrygrpolanco@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "kiwipiepy>=0.23.1",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
